Updating ESP8266 Code
Still many function prototypes left to implement. To Matt if you read
this I'm sorry for pulling a Jessie and committing broken code, but this
is the best way to tell me what I still need to do (because I can't test
all the wifi packets yet). Also adding more to the header file for msg
types to include a low battery message