Another big finish
Oh baby feels good. So the code is pretty much done on the Teensy minus
the MPU6050 integration (for sending to the ESP8266). Also we need to do
a bit more once the code to the ESP8266 is done. Motors are working well
and can be tested by sending in floats from the
testSendingFloatsFrom8266UsingTeensy.ino file